requirements.txt
about-time==4.2.1
addict==2.4.0
alive-progress==3.0.1
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
attrs==22.2.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
colorama==0.4.6
comm==0.1.2
contourpy==1.0.7
coverage==7.0.5
cycler==0.11.0
debugpy==1.6.5
decorator==4.4.2
defusedxml==0.7.1
entrypoints==0.4
exceptiongroup==1.1.0
executing==1.2.0
fastjsonschema==2.16.2
fonttools==4.38.0
fqdn==1.5.1
grapheme==0.6.0
idna==3.4
imageio==2.24.0
imageio-ffmpeg==0.4.8
imutils==0.5.4
iniconfig==2.0.0
ipykernel==6.20.1
ipython==8.8.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter_client==7.4.9
jupyter_core==5.1.3
jupyter_server==2.1.0
jupyter_server_terminals==0.4.4
jupyterlab==3.5.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
jupyterlab_server==2.18.0
kiwisolver==1.4.4
lapsolver==1.1.0
logger-tt==1.7.2
MarkupSafe==2.1.1
matplotlib==3.6.3
matplotlib-inline==0.1.6
mistune==2.0.4
mmcv-full==1.7.1
mmdet==2.27.0
moviepy==1.0.3
multiprocessing-inference @ git+https://github.com/AlphonsG/Multiprocessing-Inference@1cd3b2c086d0871fca44fc58de691f876c224ee4
natsort==8.2.0
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==3.0
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.24.1
opencv-python==4.7.0.68
packaging==23.0
pandas==1.5.2
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.6.2
pluggy==1.0.0
proglog==0.1.10
prometheus-client==0.15.0
prompt-toolkit==3.0.36
psutil==5.9.4
pure-eval==0.2.2
pycocotools==2.0.6
pycparser==2.21
pydantic==1.10.4
Pygments==2.14.0
pyparsing==3.0.9
pypeln==0.4.9
pyrsistent==0.19.3
pytest==7.2.1
pytest-cov==4.0.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-json-logger==2.0.4
pytz==2022.7.1
PyWavelets==1.4.1
pywin32==305
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.0
regex==2022.10.31
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
scikit-image==0.19.3
scipy==1.10.0
seaborn==0.12.2
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.6.2
stopit==1.1.2
terminado==0.17.1
terminaltables==3.1.10
tifffile==2022.10.10
tinycss2==1.2.1
tomli==2.0.1
torch==1.13.1+cu117
torchaudio==0.13.1+cu117
torchvision==0.14.1+cu117
tornado==6.2
tqdm==4.64.1
traitlets==5.8.1
typing_extensions==4.4.0
uri-template==1.2.0
urllib3==1.26.14
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
widgetsnbextension==4.0.5
yapf==0.32.0
